Another Light Fix...

+++ src/collision/OpticProperties.java
@@ -13,7 +13,7 @@ import game.place.Shadow;
  */
 public class OpticProperties {
 
-    public static final int FULL_SHADOW = 0, NO_SHADOW = 1, IN_SHADE_NO_SHADOW = 2, TRANSPARENT = 3, INITIAL_SHADOWS_COUNT = 20;
+    public static final byte FULL_SHADOW = 0, NO_SHADOW = 1, IN_SHADE_NO_SHADOW = 2, TRANSPARENT = 3, INITIAL_SHADOWS_COUNT = 20;
     private static final boolean[] LITABLE = {true, true, false, false};
     private static final boolean[] GIVE_SHADOW = {true, false, false, true};
     private int type;
